% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustEncoding_method.R
\docType{methods}
\name{adjustEncoding}
\alias{adjustEncoding}
\alias{adjustEncoding,character-method}
\alias{adjustEncoding,ctkPipe-method}
\title{apply iconv to files in a directory}
\usage{
adjustEncoding(.Object, ...)

\S4method{adjustEncoding}{character}(.Object, targetDir = NULL, from = NULL,
  to, xml = FALSE, ...)

\S4method{adjustEncoding}{ctkPipe}(.Object, sourceDir, targetDir,
  from = "UTF-8", to = "ISO-8859-1", xml = FALSE, log = FALSE, ...)
}
\arguments{
\item{.Object}{a character string providing a directory}

\item{targetDir}{where to put files}

\item{from}{encoding of the files in the the sourceDir, if NULL, the command line file utility will be used to detect the encoding of the source file}

\item{to}{final encoding}

\item{mc}{use multicore}

\item{progress}{whether to display progress bar}
}
\value{
something
}
\description{
apply iconv to files in a directory
}

