% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.xml_method.R
\docType{methods}
\name{as.xml}
\alias{as.xml}
\alias{as.xml,data.frame-method}
\alias{as.xml,list-method}
\title{convert data.frame to XML}
\usage{
as.xml(.Object, ...)

\S4method{as.xml}{data.frame}(.Object, meta, body, file, root = "collection",
  element = "text", clean = TRUE, verbose = TRUE)

\S4method{as.xml}{list}(.Object, filename = NULL)
}
\arguments{
\item{.Object}{object to convert (data.frame)}

\item{meta}{character string, the column names of the metadata}

\item{body}{character string, the column names of the text body, the column names will be turned into the value of a type attribute}

\item{file}{filename including the path to which the XML will be written}

\item{root}{name of the root node}

\item{element}{the name no the element to which the paragraphs will be attached}

\item{clean}{logical, if TRUE, whitespace will be removed from all columns}

\item{verbose}{logical}
}
\description{
convert data.frame to XML
}

