% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/characterCount_method.R
\docType{methods}
\name{characterCount,character-method}
\alias{characterCount,character-method}
\title{count characters in files}
\usage{
\S4method{characterCount}{character}(.Object, regexCharsToKeep = "[a-zA-Z]",
  xml = TRUE, toLower = FALSE, progress = TRUE, mc = FALSE,
  verbose = TRUE)
}
\arguments{
\item{.Object}{a directory with XML files}

\item{regexCharsToKeep}{a regex defining the characters to keep}

\item{toLower}{logical, whether to to apply tolower to string read in}

\item{progress}{defaults to TRUE}

\item{mc}{defaults to FALSE}

\item{verbose}{whether to be verbose}
}
\description{
count characters in files
}

