% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fix_method.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate vrt files for CWB import.}
\usage{
consolidate(x, sourceDir, targetDir, encoding, replacements, ...)
}
\arguments{
\item{x}{a character vector providing a directory with vrt files}

\item{sourceDir}{character vector with directory with files}

\item{targetDir}{a character vector}

\item{encoding}{encoding of the file, used by scan}

\item{replacements}{a list of character vectors (length 2 each) with regular expressions / replacements}

\item{...}{further parameters passed into \code{dirApply}}
}
\description{
Files resulting from tagging/annotation may violate the requirements of the
Corpus Workbench (CWB).  Consolidate the known issues the vrt files may cause.
}
\details{
Known issues resulting from annotating files (with the treetagger in particular)
are whitespace characters invalid for XML, XML elements at the end of a line
rather than in a seperate line, characters invalid for XML (such as ampersands),
inter alia.

Before doing respective corrections, the method tests whether there is any text at
all in the files. Empty files (files that contain nothing but XML tags) are dropped.
}

