% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreNLP_class.R
\docType{class}
\name{CoreNLP-class}
\alias{CoreNLP-class}
\alias{CoreNLP}
\title{Class for using Stanford CoreNLP}
\arguments{
\item{filename}{if filename is not NULL (default), the object will be initialized with
a FileWriter, and new annotations will be appended}

\item{colsToKeep}{character vector with names of columens of the output data.table}
}
\description{
Class for using Stanford CoreNLP
}
\section{Fields}{

\describe{
\item{\code{tagger}}{...}

\item{\code{xmlifier}}{...}

\item{\code{jsonifier}}{...}

\item{\code{writer}}{...}

\item{\code{append}}{...}

\item{\code{method}}{...}

\item{\code{colsToKeep}}{...}

\item{\code{destfile}}{...}

\item{\code{logfile}}{...}
}}


