% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cwbImport_method.R
\docType{methods}
\name{cwbImport,character-method}
\alias{cwbImport,character-method}
\title{import into CWB}
\usage{
\S4method{cwbImport}{character}(.Object, corpus = "FOO", cwbRegistry,
  sAttributes, xml = TRUE, verbose = TRUE)
}
\arguments{
\item{.Object}{a character vector}

\item{corpus}{the corpus name}

\item{cwbRegistry}{the directory of the CWB registry}

\item{sAttributes}{a list, use sAttributes-method}

\item{xml}{whether files are XML}

\item{verbose}{defaults to TRUE}
}
\description{
This is a wrapper for the cwb-encode and cwb-make command line tools.
}

