% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicates_method.R
\name{documentSimilarity}
\alias{documentSimilarity}
\title{get cosine similarity for a sparse matrix}
\usage{
documentSimilarity(x, select = NULL, method = "cosine",
  weighting = "tfidf", progress = TRUE, verbose = FALSE, mc = FALSE)
}
\arguments{
\item{x}{a simple_triplet_matrix}

\item{select}{a simple_triplet_matrix}

\item{method}{defaults to "cosine", no other method is implemented}

\item{weighting}{defaults to "tfidf"}

\item{progress}{logical}

\item{verbose}{logical}

\item{mc}{logical, or if numeric, providing number of cores}
}
\description{
get cosine similarity for a sparse matrix
}

