% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_method.R
\docType{methods}
\name{encode}
\alias{encode}
\alias{encode,character-method}
\alias{encode,ctkPipe-method}
\alias{encode,Regions-method}
\title{import into CWB}
\usage{
encode(.Object, ...)

\S4method{encode}{character}(.Object, corpus = "FOO", registry, sAttributes,
  encoding, xml = TRUE, verbose = TRUE)

\S4method{encode}{ctkPipe}(.Object, sourceDir, corpus, encoding, xml = TRUE,
  verbose = TRUE, ...)

\S4method{encode}{Regions}(.Object, filename = NULL, verbose = TRUE)
}
\arguments{
\item{.Object}{a character vector}

\item{corpus}{the corpus name}

\item{registry}{the directory of the CWB registry}

\item{sAttributes}{a list, use sAttributes-method}

\item{xml}{whether files are XML}

\item{verbose}{defaults to TRUE}

\item{corpus}{CWB corpus name that shall be created}

\item{xml}{logical}
}
\description{
This is a wrapper for the cwb-encode and cwb-make command line tools.
}
