% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findReplace_method.R
\name{findAndReplace}
\alias{findAndReplace}
\title{Find and replace.}
\usage{
findAndReplace(x, y = NULL, replacements, ...)
}
\arguments{
\item{x}{character vector either specifying a file directory, or a string to process}

\item{y}{a target directory}

\item{replacements}{a list}

\item{...}{further parameters that are passed to \code{dirApply}}
}
\description{
Find and replace.
}
\examples{
x <- c("water and milk", "butter or water", "scotch or soda")
findAndReplace(x, replacements = list(c("and", "or"), c("s", "S"), c("\\\\sw", " W")))
}
