% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findReplace_method.R
\name{findReplace}
\alias{findReplace}
\title{Find and replace.}
\usage{
findReplace(object, ...)
}
\arguments{
\item{replacements}{a list}

\item{checkValidity}{whether to validate XML in targetDir}

\item{progress}{whether to use progress bar}
}
\description{
Find and replace.
}
