% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fix_method.R
\docType{methods}
\name{fix,ctkPipe-method}
\alias{fix}
\alias{fix,ctkPipe-method}
\alias{fix-method}
\title{correct/polish vrt files}
\usage{
\S4method{fix}{ctkPipe}(x, sourceDir, targetDir, encoding = "UTF-8", ...)
}
\arguments{
\item{x}{a character vector providing a directory with vrt files}

\item{targetDir}{a character vector}

\item{mc}{whether to use multicore}

\item{verbose}{logicel, defaults to TRUE}
}
\description{
The treetagger wrapper script results in some faulty lines - this is corrected here.
}

