% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fixVrt_method.R
\docType{methods}
\name{fixVrt,character-method}
\alias{fixVrt}
\alias{fixVrt,character-method}
\alias{fixVrt-method}
\title{correct/polish vrt files}
\usage{
\S4method{fixVrt}{character}(.Object)
}
\arguments{
\item{.Object}{a character vector providing a directory with vrt files}

\item{targetDir}{a character vector}

\item{mc}{whether to use multicore}

\item{verbose}{logicel, defaults to TRUE}
}
\description{
The treetagger wrapper script results in some faulty lines - this is corrected here.
}

