% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAttributeValues_method.R
\name{getAttributeValues}
\alias{getAttributeValues}
\title{get attribute values}
\usage{
getAttributeValues(.Object, element, attrs, unique = TRUE, pattern = "xml",
  mc = FALSE, progress = TRUE, verbose = FALSE, sample = FALSE,
  filenames = NULL)
}
\arguments{
\item{.Object}{a character string providing the directory}

\item{element}{character string length 1 giving the element name}

\item{attrs}{character string providing attribute names}

\item{unique}{logical}

\item{pattern}{e.g. "xml"}

\item{mc}{logical}

\item{progress}{logical}

\item{verbose}{logical}

\item{sample}{defaults to NULL, or a numeric}

\item{filenames}{a set of filenames}
}
\value{
If attrs is a character vector of length 1, a character vector, a matrix otherwise
}
\description{
Get values of attributes of a XML node in files.
}
