% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicates_method.R
\docType{methods}
\name{getDuplicates,character-method}
\alias{getDuplicates,character-method}
\title{get duplicates}
\usage{
\S4method{getDuplicates}{character}(.Object, similarityMatrix,
  threshold = 0.9, mc = FALSE, progress = TRUE, verbose = TRUE)
}
\description{
get duplicates
}
\examples{
\dontrun{
xmlDir <- "/Users/blaette/Lab/repos/keywords/data/figaro/xml"
charCount <- characterCount(xmlDir, toLower=TRUE, progress=TRUE, verbose=TRUE, mc=3)
ngramMatrix <- getNgrams(xmlDir, charCount=noChars, nChar=10, progress=TRUE, mc=3, verbose=TRUE)
ngramMatrixWeighed <- polmineR::weigh(ngramMatrix, method="tfidf")
docsToCompareMatrix <- getPotentialDuplicates(xmlDir, reduce=TRUE, progress=TRUE, verbose=TRUE, mc=3)
similarityMatrix <- documentSimilarity(x=ngramMatrixWeighed, select=docsToCompareMatrix, verbose=TRUE, progress=TRUE, mc=3)
duplicates <- getDuplicates(xmlDir, similarityMatrix=similarityMatrix, progress=TRUE, mc=3, verbose=TRUE)
annotatedXML <- annotateDuplicates(
  .Object=xmlDir, targetDir=NULL,
  duplicates=duplicates, verbose=TRUE, mc=FALSE, progress=TRUE
  )
}
}

