% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiles_method.R
\docType{methods}
\name{getFiles}
\alias{getFiles}
\title{get files from several directories}
\usage{
\S4method{getFiles}{ctkPipe}(.Object, sourceDir, targetDir, recursive = TRUE,
  exclude = NULL, rectify = TRUE, method = "find", pattern = "html",
  verbose = T, progress = F, failsafe = FALSE)
}
\arguments{
\item{.Object}{a ctkPipe class object}

\item{sourceDir}{source directories}

\item{targetDir}{targetDir in the projectDir}

\item{recursive}{whether to check subdirectories}

\item{exclude}{I do not remember}

\item{rectify}{logical, whether to get rid of special characters in filenames}

\item{method}{either list.files or find}

\item{pattern}{the pattern for the filenames}

\item{verbose}{logical, whether to be verbose}

\item{progress}{whether to show progress bar}

\item{failsafe}{whether to be robust}
}
\description{
get files from several directories
}
