% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFiles_method.R
\name{getFiles}
\alias{.getFiles}
\alias{getFiles}
\title{Get files from several directories.}
\usage{
.getFiles(sourceDir, targetDir, recursive = TRUE, exclude = NULL,
  rectify = TRUE, method = "find", pattern = NULL, verbose = TRUE,
  progress = TRUE, failsafe = FALSE)
}
\arguments{
\item{sourceDir}{source directories}

\item{targetDir}{targetDir target directory}

\item{recursive}{whether to check subdirectories}

\item{exclude}{I do not remember}

\item{rectify}{logical, whether to get rid of special characters in filenames}

\item{method}{either list.files or find}

\item{pattern}{a pattern filenames should match, passed into \code{list.files}}

\item{verbose}{logical, whether to be verbose}

\item{progress}{whether to show progress bar}

\item{failsafe}{whether to be robust}
}
\description{
Files in the sourceDir are copied to the targetDir. The function
offers options for a robust process.
}

