% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNgrams_method.R
\docType{methods}
\name{getNgrams,character-method}
\alias{getNgrams,character-method}
\title{get ngrams}
\usage{
\S4method{getNgrams}{character}(.Object, charCount = NULL, nChar = 10,
  returnSparseMatrix = TRUE, progress = TRUE, verbose = FALSE,
  mc = FALSE)
}
\arguments{
\item{.Object}{the source directory}

\item{returnSparseMatrix}{logical, whether to return a TermDocumentMatrix, defaults to TRUE. If FALSE, a list will be returned}

\item{progress}{logical}

\item{verbose}{logical}

\item{mc}{logical, or the number of cores}

\item{.charCount}{a count of the characters in the files in the source directory}

\item{nchar}{the number of the most frequent characters that are kept}
}
\description{
get ngrams
}
\examples{
\dontrun{
xmlDir <- "/Users/blaette/Lab/repos/keywords/data/figaro/xml"
noChars <- characterCount(xmlDir, toLower=TRUE, progress=TRUE, verbose=TRUE, mc=3)
ngramMatrix <- getNgrams(xmlDir, charCount=noChars, nChar=10, progress=TRUE, mc=3, verbose=TRUE)
}
}

