% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/duplicates_method.R
\docType{methods}
\name{getPotentialDuplicates,character-method}
\alias{getPotentialDuplicates,character-method}
\title{get potential duplicates}
\usage{
\S4method{getPotentialDuplicates}{character}(.Object, dateElement = "/text",
  dateAttribute = "day", reduce = TRUE, verbose = FALSE,
  progress = TRUE, mc = FALSE)
}
\description{
get potential duplicates
}
\examples{
\dontrun{
xmlDir <- "/Users/blaette/Lab/repos/keywords/data/figaro/xml"
docsToCompareMatrix <- getPotentialDuplicates(xmlDir, reduce=TRUE, progress=TRUE, verbose=TRUE, mc=TRUE)
}
}

