% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.corenlp.R
\name{install.corenlp}
\alias{install.corenlp}
\title{Install Stanford CoreNLP.}
\usage{
install.corenlp(lang = "de")
}
\arguments{
\item{lang}{Languages to install.}
}
\description{
The function provides an installation mechanism to download and install
Stanford CoreNLP within the ctk package.
and lemmatization
}
