% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.treetagger.R
\name{install.treetagger}
\alias{install.treetagger}
\title{Install treetagger.}
\usage{
install.treetagger(lang = "de")
}
\arguments{
\item{lang}{Languages to install.}
}
\description{
The Treetagger is a standard tool for part-of-speech annotation
and lemmatization
}
