% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mergeFiles_method.R
\docType{methods}
\name{mergeFiles,ctkPipe-method}
\alias{mergeFiles,ctkPipe-method}
\title{merge files}
\usage{
\S4method{mergeFiles}{ctkPipe}(.Object, sourceDir, targetDir, regex,
  rootElement, rootAttrs, mc = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{.Object}{a ctkObject}

\item{sourceDir}{the source directory}

\item{targetDir}{the target directory}

\item{regex}{the regex to match the files}

\item{mc}{logical, whether to use parallel processing}

\item{verbose}{logical, whether to be verbose}

\item{...}{}
}
\description{
Merge files matching a regex to a new XML file.
}

