% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/missingFiles_method.R
\docType{methods}
\name{missingFiles}
\alias{missingFiles}
\alias{missingFiles,pipe-method}
\title{get files in targetDir that are missing in sourceDir}
\usage{
missingFiles(.Object, ...)

\S4method{missingFiles}{pipe}(.Object, sourceDir, targetDir)
}
\arguments{
\item{.Object}{ctkObject}

\item{...}{further parameters}

\item{sourceDir}{bla}

\item{targetDir}{bla}
}
\description{
get files in targetDir that are missing in sourceDir
}

