% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner_method.R
\docType{methods}
\name{ner}
\alias{ner}
\alias{ner,pipe-method}
\title{apply named entity recognition to files in a dir (using StanfordNLP)}
\usage{
ner(.Object, ...)

\S4method{ner}{pipe}(.Object, sourceDir, targetDir, pattern = "xml",
  mc = FALSE, progress = TRUE, verbose = FALSE, sample = FALSE,
  files = NULL, continue = FALSE, failsafe = FALSE, ...)
}
\arguments{
\item{.Object}{a ctk object}

\item{...}{further paramters}

\item{sourceDir}{the source directory}

\item{targetDir}{the target directory}

\item{pattern}{process only files matching the pattern}

\item{mc}{use multicore}

\item{progress}{show progress bar}

\item{verbose}{whether to be verbose}

\item{sample}{process only a subset}

\item{files}{process only the files specified}

\item{continue}{process only files not yet present in target dir}

\item{failsafe}{proceed in a failsafe manner}
}
\description{
apply named entity recognition to files in a dir (using StanfordNLP)
}
\examples{
\dontrun{
bt <- plprbtpdf("/home/blaette/plprbtpdf")
tokenize(bt, sourceDir="xmlCwb", targetDir="tmp")
}
}
