% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{monthsToNumeric}
\alias{monthsToNumeric}
\alias{normalizeGermanDate}
\alias{regexDate}
\title{normalize german date}
\format{An object of class \code{list} of length 12.}
\usage{
monthsToNumeric

normalizeGermanDate(dateRaw)

regexDate
}
\arguments{
\item{dateRaw}{date to process}
}
\description{
normalize german date
}
\keyword{datasets}
