% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_method.R
\docType{methods}
\name{recode}
\alias{recode}
\alias{recode,character-method}
\title{Apply iconv to files in a directory.}
\usage{
recode(.Object, ...)

\S4method{recode}{character}(.Object, targetDir = NULL, from = NULL, to,
  xml = FALSE, ...)
}
\arguments{
\item{.Object}{a character string providing a directory}

\item{targetDir}{where to put files}

\item{from}{encoding of the files in the the sourceDir, if NULL, the command line file utility will be used to detect the encoding of the source file}

\item{to}{final encoding}

\item{mc}{use multicore}

\item{progress}{whether to display progress bar}
}
\value{
something
}
\description{
Apply iconv to files in a directory.
}
