% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regex_method.R
\docType{methods}
\name{regex,character-method}
\alias{regex,character-method}
\title{apply regex to a set of files}
\usage{
\S4method{regex}{character}(object, subdir = NULL, pattern = ".*\\\\.xml",
  regex = "<.*?>")
}
\arguments{
\item{object}{a path}

\item{subdir}{provide a subdirectory}

\item{pattern}{what to find}

\item{regex}{regex to be applied}
}
\description{
apply regex to a set of files
}

