% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regexContext_function.R
\name{regexContext}
\alias{regexContext}
\title{get context of a regex}
\usage{
regexContext(.Object, ...)
}
\arguments{
\item{.Object}{bla}

\item{filenames}{defaults to NULL, then all files will be examined, of provided, only specific files}

\item{regex}{a regex}

\item{ncharContext}{characters to the left and to the right}

\item{verbose}{logical}

\item{progress}{logical}

\item{mc}{logical or numeric}
}
\description{
Get a data frame with the left and right context of the match of a regex.
}

