% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{removeWhitespace}
\alias{removeWhitespace}
\title{Remove whitespace.}
\usage{
removeWhitespace(x, rmBlankLines = TRUE)
}
\arguments{
\item{x}{a character vector to process}

\item{rmBlankLines}{logical, whether to remove blank lines}
}
\description{
Remove leading and trailing whitespace from a character vector.
}

