% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceInvalidCharacters_function.R
\docType{methods}
\name{replaceInvalidCharacters}
\alias{replaceInvalidCharacters}
\title{replace invalid characters}
\usage{
\S4method{replaceInvalidCharacters}{ctkPipe}(.Object, sourceDir, targetDir,
  xml = FALSE, ...)
}
\arguments{
\item{sourceDir}{directory with files to be processed}

\item{targetDir}{defaults to NULL, if provided, the directory where the processed files are saved}

\item{xml}{defaults to FALSE, if TRUE, files are provessed as XML files (not yet implemented)}

\item{txt}{defaults to NULL, if provided, a text string to be processed}

\item{filenames}{defaults to NULL, if provided, a set of files in the sourceDir}

\item{mc}{logcial or numeric, whether to use multicore and the number of cores}

\item{progress}{logical}

\item{verbose}{logical}
}
\description{
Replace characters that do not conform to UTF-8 encoding. If txt is NULL, all files
in the source dir are processed.
}
