% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAttributeList_method.R
\name{sAttributeList}
\alias{sAttributeList}
\title{get a list with sAttributes from files}
\usage{
sAttributeList(.Object, sample = 100, ...)
}
\arguments{
\item{.Object}{a directory}

\item{sample}{numeric, a subsample of files to derive the s-attributes from}

\item{...}{further parameters that are passed into \code{dirApply}}
}
\description{
get a list with sAttributes from files
}
