% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.saxon.R
\name{install.saxon}
\alias{install.saxon}
\title{Install Saxon XSLT Processor to tools directory.}
\usage{
install.saxon()
}
\description{
The Saxon parser offers functionality to perform XSL
  transformations efficiently. It is used by the \code{xslt} function, which
  calls Saxon. This function installs the Saxon parser to the subdirectory
  tools/saxon of the ctk package directory.
}
