% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setupPipeDir_method.R
\name{setupPipeDir}
\alias{setupPipeDir}
\title{Setup pipeDir.}
\usage{
setupPipeDir(.Object, ...)
}
\arguments{
\item{.Object}{the object}

\item{sub}{the subdirectories to create (character vector)}

\item{main}{the main directory}
}
\description{
Create pipeDir and all relevant subdirectories.
}

