% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_method.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize files.}
\usage{
tokenize(.Object, ...)
}
\arguments{
\item{.Object}{a ctk object}

\item{...}{further paramters}

\item{sourceDir}{the source directory}

\item{targetDir}{the target directory}

\item{with}{either "stanfordNLP", "treetagger" or "openNLP"}

\item{lang}{language of the files to be tagged}

\item{pattern}{process only files matching the pattern}

\item{mc}{use multicore}

\item{progress}{show progress bar}

\item{verbose}{whether to be verbose}

\item{sample}{process only a subset}

\item{files}{process only the files specified}

\item{continue}{process only files not yet present in target dir}

\item{failsafe}{proceed in a failsafe manner}
}
\description{
Tokenize (XML) files with one standard tool (treetagger, stanfordNLP, openNLP).
}
\details{
One potential problem with the perl-tokenizer that comes with the treetagger
is that the output is not valid XML. It is necessary to fix the XML with a 
shell command such as \code{for i in $(ls); do sed 's/\xC2\xA0/ /g' $i > ../tok2/$i; done}.
The XML may still not be valid ("&" etc.), so fix method is still necessary.
}
