% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_method.R
\docType{methods}
\name{tokenize}
\alias{tokenize}
\alias{tokenize,character-method}
\title{Tokenize files.}
\usage{
tokenize(.Object, ...)

\S4method{tokenize}{character}(.Object, lang = "de", with = "stanfordNLP",
  ...)
}
\arguments{
\item{.Object}{a ctk object}

\item{...}{further paramters}

\item{lang}{language of the files to be tagged}

\item{with}{either "stanfordNLP", "treetagger" or "openNLP"}
}
\description{
Tokenize (XML) files with one standard tool (treetagger, stanfordNLP, openNLP).
}
\details{
One potential problem with the perl-tokenizer that comes with the treetagger
is that the output is not valid XML. It is necessary to fix the XML with a 
shell command such as \code{for i in $(ls); do sed 's/\xC2\xA0/ /g' $i > ../tok2/$i; done}.
The XML may still not be valid ("&" etc.), so fix method is still necessary.
}
