% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizeSentences_method.R
\name{tokenizeSentences}
\alias{tokenizeSentences}
\title{Tokenize sentences.}
\usage{
tokenizeSentences(.Object, ...)
}
\arguments{
\item{targetElement}{where to finde the text nodes}

\item{para}{logical, whether to annotate paragraphs}
}
\value{
the output of the procedure
}
\description{
Tokenize sentences.
}
\author{
Andreas Blaette
}
