% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetagger_method.R
\name{treetagger}
\alias{treetagger}
\alias{.treetagger}
\title{use TreeTagger}
\usage{
.treetagger(filename, sourceDir = NULL, targetDir = NULL, verbose = FALSE,
  param = list())
}
\arguments{
\item{targetDir}{output directory}

\item{verbose}{logical, defaults to TRUE}

\item{.Object}{a path}

\item{lang}{the language to be used (defaults to 'de')}

\item{treetaggerPath}{directory with the treetagger}

\item{parallel}{whether to use parallel computing}

\item{lang}{the language to use}
}
\description{
use TreeTagger
}
