% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treetagger_method.R
\name{treetagger}
\alias{treetagger}
\title{Use TreeTagger for linguistic annotation.}
\usage{
treetagger(filename, sourceDir = NULL, targetDir = NULL, verbose = FALSE,
  param = list(lang = "de", tokenize = TRUE))
}
\arguments{
\item{filename}{file to process, or a character vector (if sourceDir is NULL)}

\item{targetDir}{output directory, if NULL, the processed input will be returned}

\item{verbose}{logical, defaults to TRUE}

\item{param}{a list that needs to include the language to be used (defaults to 'de') and a logical
vector \code{tokenize} whether the input needs to be tokenized before tagging}
}
\description{
The argument \code{param} is a list passing the arguments \code{lang}, a 
character vector that is expected to be "de", "fr", "it", or "en". The argument
\code{tokenize} is a logical value. If \code{TRUE}, the tokenizer included in
the treetagger scripts used, if \code{FALSE}, the input is expected to be
tokenized already.
}
\details{
Depending whether targetDir is defined (i.e. not \code{NULL}), output is written to the
file, or a character vector is returned. If sourceDir is \code{NULL}, filename will serve
as the input character string. It will be written to a temporary file for further
processing.
}
