% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_method.R
\name{validate}
\alias{validate}
\title{Validate XML files.}
\usage{
validate(x, targetDir = NULL, dtd = NULL, ...)
}
\arguments{
\item{x}{character vector, a filename, or a directory path}

\item{targetDir}{the target directory}

\item{dtd}{filename of a DTD}

\item{...}{arguments that will be passed into dirApply}
}
\description{
A single file, or all files in directory x are checked for
  validity using the command line tool \code{xmllint}. If \code{targetDir} is
  specified, the option '--recover' will be used, and repaired XML files are
  written to targetDir. If the param \code{dtd} provides a filename with a
  DTD, files are checked against the respective DTD.
}
\examples{
\dontrun{
  validate("/Users/blaette/Lab/repos/plprnwhtm/html/2015-02-23/utf8", dtd=NULL)
  for i in $(ls); do xmllint --recover $i -o ../xmlValidated/$i; done
}
}
