% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlToDT_function.R
\name{xmlToDT}
\alias{xmlToDT}
\title{Parse an XML/TEI file as data.table}
\usage{
xmlToDT(filename, body = "//body", meta = NULL, mc = NULL)
}
\arguments{
\item{filename}{the name of the file to read in}

\item{body}{an xpath expression defining the body of the xml document}

\item{meta}{a named character vector with xpath expressions}

\item{mc}{a numeric/integer value, will be passed to param \code{mc.cores}
in a call to \code{mclapply} - number of cores to use}
}
\description{
The procedure is inspired by the tidytext paradigm, but
uses a data.table for efficiency.
}
