% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslt_method.R
\name{xslt}
\alias{xslt}
\title{perform XSL transformation}
\usage{
xslt(.Object, ...)
}
\arguments{
\item{.Object}{a directory}

\item{targetDir}{output directory}

\item{xslFile}{filename}

\item{mc}{logical, whether to use multicore}

\item{verbose}{whether to be verbose, defaults to TRUE}
}
\description{
perform XSL transformation
}

