% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xslt_method.R
\name{xslt}
\alias{xslt}
\title{Perform XSL transformation.}
\usage{
xslt(filename, sourceDir = NULL, targetDir = NULL, verbose = FALSE,
  param = list(xslFile = character()))
}
\arguments{
\item{filename}{a XML file to be processed}

\item{sourceDir}{directory where the XML file resides}

\item{targetDir}{output directory}

\item{verbose}{logical value, whether to be verbose}

\item{param}{named list of further parameters, it needs to include xslFile,
the filename of the XSL file}
}
\value{
return of the saxon parser
}
\description{
Worker function to be iterated over all files using \code{dirApply}.
}
