% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.package_function.R
\name{as.package}
\alias{as.package}
\alias{unpack}
\title{create R data package with corpus}
\usage{
as.package(corpus, targetDir, author, pkg = NULL, version = NULL,
  date = NULL, maintainer = "Andreas Blaette <andreas.blaette@uni-due.de>",
  description = NULL, license = "NONE", verbose = TRUE)

unpack(pkg, verbose = TRUE)
}
\arguments{
\item{corpus}{a corpus detailes in CORPUS_REGISTRY dir}

\item{targetDir}{the target directory}

\item{author}{the author of the package}

\item{pkg}{package name}

\item{version}{the version number of the corpus}

\item{date}{date of creation}

\item{maintainer}{maintainer, R package style}

\item{description}{short description of the data package}

\item{verbose}{whether to be verbose}

\item{licence}{the license}
}
\description{
create R data package with corpus
}
\examples{
\dontrun{
as.package("ARENEN", targetDir="/home/blaette/cwb.arenen", author="Andreas Blaette")
}
}

