% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/checkDirs_method.R, R/ctkPipe_class.R
\docType{methods}
\name{checkDirs,ctkPipe-method}
\alias{checkDirs,ctkPipe-method}
\alias{ctkPipe-class}
\title{check whether necessary directories exist}
\usage{
\S4method{checkDirs}{ctkPipe}(.Object, sourceDir, targetDir)
}
\arguments{
\item{.Object}{}

\item{sourceDir}{directory1}

\item{targetDir}{}
}
\description{
check whether necessary directories exist

ctkPipe class
}
\section{Slots}{

\describe{
\item{\code{projectDir}}{a character string}

\item{\code{remoteDir}}{dir on server}

\item{\code{subdirs}}{character}

\item{\code{summary}}{a data.frame}

\item{\code{call}}{the calls that have been applied}

\item{\code{saxonPath}}{path to saxon}

\item{\code{treetaggerPath}}{path to treetagger}

\item{\code{cwbRegistry}}{dir of cwb-registry}

\item{\code{tasks}}{what to do (data.frame)}

\item{\code{sAttributes}}{a list generated by the method sAttributesList}
}}

