% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/encode_method.R
\docType{methods}
\name{encode}
\alias{encode}
\alias{encode,character-method}
\alias{encode,regions-method}
\title{Encode corpus (CWB import).}
\usage{
encode(.Object, ...)

\S4method{encode}{character}(.Object, corpus = "FOO",
  registry = Sys.getenv("CORPUS_REGISTRY"), sAttributes, embedding = "0",
  encoding = "utf8", xml = TRUE, exec = TRUE, verbose = TRUE, ...)

\S4method{encode}{regions}(.Object, filename = NULL, verbose = TRUE)
}
\arguments{
\item{.Object}{a character vector}

\item{...}{further parameters (not passed)}

\item{corpus}{the corpus name}

\item{registry}{the directory of the CWB registry}

\item{sAttributes}{a list, use sAttributes-method}

\item{embedding}{levels of embedding elements to allow for}

\item{encoding}{encoding, passed into \code{cwb-encode}}

\item{xml}{whether files are XML}

\item{exec}{logical, whether to execute commands}

\item{verbose}{defaults to TRUE}

\item{filename}{a filename}
}
\description{
This is a wrapper for the \code{cwb-encode} and \code{cwb-make} command line tools.
The calls are generated and executed (via a system call).
}

