% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_method.R
\docType{methods}
\name{encode}
\alias{encode}
\alias{encode,Regions-method}
\alias{encode,character-method}
\alias{encode,pipe-method}
\title{Encode corpus (CWB import).}
\usage{
encode(.Object, ...)

\S4method{encode}{character}(.Object, corpus = "FOO",
  registry = Sys.getenv("CORPUS_REGISTRY"), sAttributes, embedding = "0",
  encoding = "utf8", xml = TRUE, exec = TRUE, verbose = TRUE, ...)

\S4method{encode}{pipe}(.Object, sourceDir, corpus, encoding, xml = TRUE,
  verbose = TRUE, embedding = "0", exec = TRUE, ...)

\S4method{encode}{Regions}(.Object, filename = NULL, verbose = TRUE)
}
\arguments{
\item{.Object}{a character vector}

\item{corpus}{the corpus name}

\item{registry}{the directory of the CWB registry}

\item{sAttributes}{a list, use sAttributes-method}

\item{embedding}{levels of embedding elements to allow for}

\item{xml}{whether files are XML}

\item{exec}{logical, whether to execute commands}

\item{verbose}{defaults to TRUE}

\item{corpus}{CWB corpus name that shall be created}

\item{xml}{logical}
}
\description{
This is a wrapper for the \code{cwb-encode} and \code{cwb-make} command line tools. 
The calls are generated and executed (via a system call).
}

