% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/missingFiles_method.R
\docType{methods}
\name{missingFiles}
\alias{missingFiles}
\alias{missingFiles,ctkPipe-method}
\title{get files in targetDir that are missing in sourceDir}
\usage{
missingFiles(.Object, ...)

\S4method{missingFiles}{ctkPipe}(.Object, sourceDir, targetDir)
}
\arguments{
\item{.Object}{ctkObject}

\item{...}{}

\item{sourceDir}{bla}

\item{targetDir}{bla}
}
\description{
get files in targetDir that are missing in sourceDir
}

