% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{monthsToNumeric}
\alias{monthsToNumeric}
\alias{normalizeGermanDate}
\alias{regexDate}
\title{normalize german date}
\format{\preformatted{List of 12
 $ Januar   : chr "01"
 $ Februar  : chr "02"
 $ März     : chr "03"
 $ April    : chr "04"
 $ Mai      : chr "05"
 $ Juni     : chr "06"
 $ Juli     : chr "07"
 $ August   : chr "08"
 $ September: chr "09"
 $ Oktober  : chr "10"
 $ November : chr "11"
 $ Dezember : chr "12"
}}
\usage{
monthsToNumeric

normalizeGermanDate(dateRaw)

regexDate
}
\arguments{
\item{dateRaw}{date to process}
}
\description{
normalize german date
}
\keyword{datasets}

