% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkDirs_method.R, R/pipe_class.R
\docType{methods}
\name{checkDirs,pipe-method}
\alias{checkDirs,pipe-method}
\alias{pipe-class}
\title{check whether necessary directories exist}
\usage{
\S4method{checkDirs}{pipe}(.Object, sourceDir, targetDir)
}
\arguments{
\item{.Object}{object}

\item{sourceDir}{directory1}

\item{targetDir}{target directory}
}
\description{
check whether necessary directories exist

pipe class
}
\section{Slots}{

\describe{
\item{\code{projectDir}}{a character string}

\item{\code{remoteDir}}{dir on server}

\item{\code{subdirs}}{character}

\item{\code{summary}}{a data.frame}

\item{\code{call}}{the calls that have been applied}

\item{\code{saxonPath}}{path to saxon}

\item{\code{treetaggerPath}}{path to treetagger}

\item{\code{registry}}{dir of cwb-registry}

\item{\code{tasks}}{what to do (data.frame)}

\item{\code{sAttributes}}{a list generated by the method sAttributesList}
}}

