% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeDir_class.R
\docType{class}
\name{PipeDir-class}
\alias{PipeDir-class}
\alias{Pipe}
\title{Pipe for corpus preparation.}
\arguments{
\item{sourceDir}{a subdirectory of the pipeDir where files to be processed
are located}

\item{targetDir}{a subdirectory of the pipeDir where processed output is
stored}
}
\description{
Pipe for corpus preparation.
}
\section{Fields}{

\describe{
\item{\code{basetable}}{a \code{data.table} that includes metadata and a column
'text' with the full text of the corpus}

\item{\code{metadata}}{a \code{data.table} with the metadata for the corpus; a
column 'chunk' serves as an id to link the metadata table with the other
tables for corpus preparation}

\item{\code{texttable}}{a \code{data.table} derived from the basetable, includes
only the column 'chunk' with the id, and the column 'text' with the
unparsed full text}

\item{\code{tokenstream}}{a \code{table} with a columns with the tokenized text and
further annotation of the corpus}
}}

\section{Methods}{

\describe{
\item{\code{corenlp(sourceDir = "csv", targetDir = "ndjson")}}{Use Stanford CoreNLP.}

\item{\code{ndjsonToDf(sourceDir = "ndjson", targetDir = "csv")}}{Convert ndjson resulting from Stanford to csv.}

\item{\code{xmlToBasetable(sourceDir = "xml", targetDir = "csv", metadata)}}{Extract text and metadata from XML documents, and keep the resulting
'basetable' in the respective slot of the Pipe object.}
}}

