% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_method.R
\docType{methods}
\name{regex,character-method}
\alias{regex,character-method}
\title{Get Matches for Regular Expression in Files in Directory.}
\usage{
\S4method{regex}{character}(.Object, pattern = ".*\\\\.xml",
  regex = "<.*?>")
}
\arguments{
\item{.Object}{a directory}

\item{pattern}{a regular expression passed into \code{list.files} to filter files}
}
\description{
Get Matches for Regular Expression in Files in Directory.
}
