% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeWhitespace}
\alias{removeWhitespace}
\title{remove leading and trailing whitespace of a character vector}
\usage{
removeWhitespace(x, rmBlankLines = TRUE)
}
\description{
remove leading and trailing whitespace of a character vector
}

