% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAttributeList_method.R
\docType{methods}
\name{sAttributeList,character-method}
\alias{sAttributeList,character-method}
\title{get a list with sAttributes from files}
\usage{
\S4method{sAttributeList}{character}(.Object, ...)
}
\arguments{
\item{object}{a directory}

\item{mc}{whether to use multicore}
}
\description{
get a list with sAttributes from files
}

