% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctk_package.R
\name{timePerFile-class}
\alias{timePerFile-class}
\alias{timePerFile_class}
\alias{print.timePerFile}
\title{Time per file.}
\usage{
\method{print}{timePerFile}(x, ...)
}
\arguments{
\item{x}{object to show/pring}

\item{...}{further parameters (not used)}
}
\description{
Class and methods to report time consumed to process
files.
}
