% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tokenize_method.R
\docType{methods}
\name{tokenize}
\alias{tokenize}
\alias{tokenize,ctkPipe-method}
\title{tokenize files in a dir (using StanfordNLP)}
\usage{
tokenize(.Object, ...)

\S4method{tokenize}{ctkPipe}(.Object, sourceDir, targetDir,
  with = "stanfordNLP", lang = "de", ...)
}
\arguments{
\item{.Object}{a ctk object}

\item{...}{further paramters}

\item{sourceDir}{the source directory}

\item{targetDir}{the target directory}

\item{with}{either "stanfordNLP", "treetagger" or "openNLP"}

\item{lang}{language of the files to be tagged}

\item{pattern}{process only files matching the pattern}

\item{mc}{use multicore}

\item{progress}{show progress bar}

\item{verbose}{whether to be verbose}

\item{sample}{process only a subset}

\item{files}{process only the files specified}

\item{continue}{process only files not yet present in target dir}

\item{failsafe}{proceed in a failsafe manner}
}
\description{
tokenize files in a dir (using StanfordNLP)
}

