% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizeSentences_method.R
\docType{methods}
\name{tokenizeSentences}
\alias{tokenizeSentences}
\alias{tokenizeSentences,character-method}
\title{Tokenize sentences.}
\usage{
tokenizeSentences(.Object, ...)

\S4method{tokenizeSentences}{character}(.Object, targetDir,
  targetElement = "p", para = FALSE, ...)
}
\arguments{
\item{.Object}{bla}

\item{...}{further parameters that are passed into \code{dirApply}}

\item{targetDir}{the target directory}

\item{targetElement}{where to finde the text nodes}

\item{para}{logical, whether to annotate paragraphs}
}
\value{
the output of the procedure
}
\description{
Tokenize sentences.
}
\author{
Andreas Blaette
}
