% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/treetagger_method.R
\name{.treetaggerWorker}
\alias{.treetaggerWorker}
\title{use TreeTagger}
\usage{
.treetaggerWorker(filename, sourceDir = NULL, targetDir = NULL,
  verbose = FALSE, param = list())
}
\arguments{
\item{targetDir}{output directory}

\item{verbose}{logical, defaults to TRUE}

\item{.Object}{a path}

\item{treetaggerPath}{directory with the treetagger}

\item{parallel}{whether to use parallel computing}

\item{lang}{the language to use}
}
\description{
use TreeTagger
}

