% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_method.R
\docType{methods}
\name{validate}
\alias{validate}
\alias{validate,character-method}
\alias{validate,pipe-method}
\title{validate XML files}
\usage{
validate(.Object, ...)

\S4method{validate}{character}(.Object, dtd = NULL, targetDir = NULL,
  verbose = FALSE, mc = FALSE, progress = TRUE)

\S4method{validate}{pipe}(.Object, sourceDir, targetDir = NULL, dtd = NULL,
  verbose = FALSE, mc = FALSE, progress = TRUE, files = NULL,
  pattern = "xml")
}
\arguments{
\item{.Object}{either a character vector indicating a directore, or a pipe object}

\item{...}{further arguments}

\item{dtd}{check agains a dtd}

\item{sourceDir}{the source directory}
}
\description{
a story to be told
}
\examples{
\dontrun{
  validate("/Users/blaette/Lab/repos/plprnwhtm/html/2015-02-23/utf8", dtd=NULL)
  for i in $(ls); do xmllint --recover $i -o ../xmlValidated/$i; done
}
}

